New Tiling Scheme
After discovering gitHub imposes bandwidth limits when using Git LFS, I
updated this KMZ to reference a folder of small KMZ files hosted in this
repository.  This smaller KMZ files used to be embedded in HYDE.kmz.
Now that it is not, the initial download is much faster and the
individual tiles are only downloaded as needed.